// types/applicant-upload.ts
export type ApplicantUploadRow = {
  id_tkm: string;
  nama_group: string;
  tanggal_daftar: string; // atau Date jika diubah
  role: string; // bisa diubah ke ApplicantRole jika perlu
  email: string;
  nama: string;
  nik: string;
  tgl_lahir: string;
  umur: number;
  gender: string;
  aktivitas_saat_ini: string;
  link_kk: string;
  no_kk: string;
  pendidikan_terakhir: string;
  alamat_ktp: string;
  provinsi_ktp: string;
  kota_ktp: string;
  kecamatan_ktp: string;
  desa_ktp: string;
  kode_pos_ktp: string;
  alamat_domisili: string;
  provinsi_domisili: string;
  kota_domisili: string;
  kecamatan_domisili: string;
  kelurahan_domisili: string;
  kode_pos_domisili: string;
  link_map_domisili: string;

  penyandang_disabilitas: string;
  jenis_disabilitas: string;
  link_surat_disabilitas: string;
  link_ktp: string;
  link_pas_foto: string;
  jenis_medsos: string;
  nama_medsos: string;
  no_telp: string;
  no_whatsapp: string;
  nama_kerabat_1: string;
  no_kerabat_1: string;
  status_kerabat_1: string;
  nama_kerabat_2: string;
  no_kerabat_2: string;
  status_kerabat_2: string;
  nib: string;
  nama_usaha: string;
  sektor_usaha: string;
  jenis_usaha: string;
  deskripsi_usaha: string;
  produk_utama: string;
  tipe_lokasi: string;
  status_lokas: string;
  alamat_usaha: string;
  provinsi_usaha: string;
  kota_usaha: string;
  kecamatan_usaha: string;
  kelurahan_usaha: string;
  kode_pos_usaha: string;
  link_map_usaha: string;
  saluran_pemasaran: string;
  wilayah_pemasaran: string;
  negara_pemasaran: string;
  partner: string;
  jumlah_partner: string;
  omzet: string;
  profit: string;
  unit_terjual: string;
  dokumen1_legalitas_type: string;
  dokumen1_legalitas_nama: string;
  dokumen1_legalitas_nomor: string;
  dokumen1_legalitas_link: string;
  dokumen2_legalitas_type: string;
  dokumen2_legalitas_nama: string;
  dokumen2_legalitas_nomor: string;
  dokumen2_legalitas_link: string;

  tenaga_kerja1_nama: string;
  tenaga_kerja1_nik: string;
  tenaga_kerja1_ktp: string;
  tenaga_kerja1_status: string;
  tenaga_kerja1_tipe: string;
  tenaga_kerja2_nama: string;
  tenaga_kerja2_nik: string;
  tenaga_kerja2_ktp: string;
  tenaga_kerja2_status: string;
  tenaga_kerja2_tipe: string;

  link_video_usaha: string;
  link_foto_usaha: string;

  link_proposal: string;
  link_surat_kesanggupan: string;
  lpj_outline_laporan_pertanggung_jawaban_teknis: string;
  lpj_outline_laporan_pertanggung_jawaban_keuangan: string;
  lpj_berita_acara_serah_terima_penyelesaian_pekerjaan: string;
  link_recap_laporan_keuangan_pdf: string;
  link_recap_laporan_teknis_pdf: string;
  bukti_lpj_bantuan_sebelumnya: string;
  bukti_surat_kelompok_sebelumnya: string;
  link_keterangan_afirmasi: string;
  mou_mou_user: string;
  mou_mou_admin: string;
  mou_bast_user: string;
  link_surat_pelatihan: string;
  link_surat_keterangan_desmigratif: string;
  link_dokumen_persyaratan: string;
  lpj_laporan: string;
  lpj_bukti_pengeluaran: string;
  lpj_bukti_sektor_sisa_dana: string;
};